Improve python package
This fixes python packaging by having files organized more in a way
it is supposed to be in python.
